Update - v1.03 - 27/03/2025
1- The login function is operational, and Bcrypt has been successfully tested.
2- A drop-down menu has been implemented for logged-in users, allowing them to edit their profiles and log out.
3- Additional comments and organizational improvements have been made to the existing code.

// ===================================== //

Next steps:

a) Develop a profile page where users can edit their avatar, first and last name, address, password, and more.
b) Improve the styling of the profile drop-down menu.
c) Add a default image for the avatar and store it in the database.
d) Begin creating the registration function.